Fiducial model: South Pole Telescope (SPT)
The cosmological parameters and the configuration of the survey (area, mass and redshift ranges)
were modified (in both scripts) following arXiv:1203.5775, where we consider the full SPT area (2500 deg2).

In NCountSimul.py:
The multiplicity function to be used is Nc...TinkerCrit, with Delta = 500.
The standard deviation of the ln-normal mass distribution is 0.24 (similar to those from SPT mass-observable relation).

In sequential_ABC.py:
Values of dm (mass bin) were modified.

New files are equivalent to NCountSimul.py (ln-normal mass distribution), but containing different mass-observable relations -
Nc...Benson (SPT) and Nc...Nodist (where we assume that the true masses are known).